requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-countries==7.3.1
django-currentuser==0.5.3
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-restql==0.15.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pygraphviz==1.9
PyJWT==2.3.0
pyparsing==3.0.7
pyPEG2==2.15.2
python-magic==0.4.25
python-memcached==1.59
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
